Semester 1/Hackathon 1/app/requirements.txt
streamlit

numpy==1.24.0
pandas
python-dotenv

matplotlib==3.7.2
altair
pydeck

scikit-image
imgviz

tensorflow==2.11
ultralytics

pydub
librosa